Teach find_button() to find elements with attribute of role="button"
This is in addition to the other buttons it already knows how to find,
(such as <input> and <button>). The new code will match against any
element that has a role="button" attribute.

This commit includes a new gamut of tests to exercise a variety of
different, supported button types.